INT-466: Second attempt at getting CI working
It seems that running `npm publish` from inside `yarn publish` does not
work. Possibly because of environment variables.

This PR separates out the build step so yarn isn't involved in the
publish step.

Additionally it ensures the version number of the build being published
is set to the same as the main package.json file so it is impossible to
accidently miss updating it as was a problem before.